Edited

+++ server/app.py
@@ -1,16 +1,15 @@
 #!/usr/bin/env python3
 
-from flask import Flask, request, make_response
+from flask import Flask, jsonify, request, make_response
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
-from flask import jsonify
 
 from models import db, User, Review, Game
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///app.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.json.compact = False
+app.config['JSONIFY_PRETTYPRINT_REGULAR'] = True
 
 migrate = Migrate(app, db)
 
@@ -25,16 +24,11 @@ def games():
 
     games = []
     for game in Game.query.all():
-        game_dict = {
-            "title": game.title,
-            "genre": game.genre,
-            "platform": game.platform,
-            "price": game.price,
-        }
+        game_dict = game.to_dict()
         games.append(game_dict)
 
     response = make_response(
-        games,
+        jsonify(games),
         200
     )
 
@@ -42,31 +36,33 @@ def games():
 
 @app.route('/games/<int:id>')
 def game_by_id(id):
-    game = Game.query.filter(Game.id == id).first()
-    
+    game = Game.query.filter_by(id=id).first()
+
     game_dict = game.to_dict()
 
     response = make_response(
-        game_dict,
+        jsonify(game_dict),
         200
     )
 
     return response
 
-@app.route('/reviews/<int:id>', methods=['GET', 'POST', 'PATCH', 'DELETE'])
-def review_by_id(id):
-    review = Review.query.filter_by(id=id).first()
+@app.route('/reviews', methods=['GET', 'POST'])
+def reviews():
 
     if request.method == 'GET':
-        review_dict = review.to_dict()
+        reviews = []
+        for review in Review.query.all():
+            review_dict = review.to_dict()
+            reviews.append(review_dict)
 
         response = make_response(
-            jsonify(review_dict),
+            jsonify(reviews),
             200
         )
 
         return response
-    
+
     elif request.method == 'POST':
         new_review = Review(
             score=request.form.get("score"),
@@ -86,40 +82,63 @@ def review_by_id(id):
         )
 
         return response
-       
-    elif request.method == 'PATCH':
-        review = Review.query.filter_by(id=id).first()
 
-        for attr in request.form:
-            setattr(review, attr, request.form.get(attr))
+@app.route('/reviews/<int:id>', methods=['GET', 'PATCH', 'DELETE'])
+def review_by_id(id):
+    review = Review.query.filter_by(id=id).first()
 
-        db.session.add(review)
-        db.session.commit()
-
-        review_dict = review.to_dict()
-
-        response = make_response(
-            jsonify(review_dict),
-            200
-        )
-
-        return response
-
-    elif request.method == 'DELETE':
-        db.session.delete(review)
-        db.session.commit()
-
+    if review == None:
         response_body = {
-            "delete_successful": True,
-            "message": "Review deleted."    
+            "message": "This record does not exist in our database. Please try again."
         }
-
-        response = make_response(
-            jsonify(response_body),
-            200
-        )
+        response = make_response(jsonify(response_body), 404)
 
         return response
+
+    else:
+        if request.method == 'GET':
+            review_dict = review.to_dict()
+
+            response = make_response(
+                jsonify(review_dict),
+                200
+            )
+
+            return response
+
+        elif request.method == 'PATCH':
+            review = Review.query.filter_by(id=id).first()
+
+            for attr in request.form:
+                setattr(review, attr, request.form.get(attr))
+
+            db.session.add(review)
+            db.session.commit()
+
+            review_dict = review.to_dict()
+
+            response = make_response(
+                jsonify(review_dict),
+                200
+            )
+
+            return response
+
+        elif request.method == 'DELETE':
+            db.session.delete(review)
+            db.session.commit()
+
+            response_body = {
+                "delete_successful": True,
+                "message": "Review deleted."    
+            }
+
+            response = make_response(
+                jsonify(response_body),
+                200
+            )
+
+            return response
 
 @app.route('/users')
 def users():
@@ -130,11 +149,11 @@ def users():
         users.append(user_dict)
 
     response = make_response(
-        users,
+        jsonify(users),
         200
     )
 
     return response
 
 if __name__ == '__main__':
-    app.run(port=5555, debug=True)
+    app.run(port=5555)
